Add default seedList.seed
When releasing a game with PopcornFX, you need to include these PopcornFX assets in your asset bundles (and the assets for any FX you use).  Every gem can provide a default seed list in 'Assets/seedList.seed' and it will be automatically listed in the Asset Bundler.  I pulled this list from MultiplayerSample 'BasePopcornFXSeedList.seed' and added the .varianttree files.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>